using Flask's Blueprint for weather endpoints
